Create guessthenumber.py
The main objective of the project is to develop a computer application that requires the user to predict the number after a random number is chosen from a range. The user has three attempts to estimate the number; if he gets it right, a message indicating "you guessed correctly" appears; otherwise, a message to the contrary appears.